+++ ui-components/index.ts
@@ -120,6 +120,7 @@ export * from "./src/page_components/listing/listing_sidebar/events/LotteryResul
 export * from "./src/page_components/listing/listing_sidebar/events/PublicLotteryEvent"
 export * from "./src/page_components/listing/listing_sidebar/events/EventDateSection"
 export * from "./src/page_components/sign-in/FormSignIn"
+export * from "./src/page_components/forgot-password/FormForgotPassword"
 
 /* Responsive Wrappers */
 export * from "./src/sections/ResponsiveWrappers"